move: other ui components as entities layer
