+++ apps/web/src/entities/exercise/ui/ExerciseDataList.stories.tsx
