+++ apps/web/src/entities/exercise/ui/ExerciseDataList.tsx
@@ -1,8 +1,8 @@
 import type { Schedule } from 'fitness-struct';
 import { Suspense } from 'react';
 import { useGetExerciseListByScheduleId } from '@entities/exercise/api';
+import SetListEditor from '@entities/set/ui/SetListEditor';
 import MenuableAccordion from '@shared/ui/MenuableAccordion';
-import SetListEditor from '../Sets/SetListEditor';
 
 export interface ExerciseDataListProps {
   schedule: Schedule.Schedule
