+++ apps/web/src/entities/exercisePreset/ui/PresetDisplay.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import { useExercisePresetWithListFragment } from '@entities/exercisePreset/api';
 import { usePresetMenu } from '@entities/exercisePreset/hooks';
 import MenuableAccordion from '@shared/ui/MenuableAccordion';
-import SimpleFitnessList from '@ui/Fitness/SimpleFitnessList';
+import SimpleFitnessList from '@entities/fitness/ui/SimpleFitnessList';
 
 interface PresetDisplayProps {
   presetId: number
