+++ apps/web/src/entities/fitness/ui/FitnessItem.stories.tsx
