+++ apps/web/src/entities/fitness/ui/FitnessItem.tsx
