+++ apps/web/src/entities/fitness/ui/FitnessList.stories.tsx
