+++ apps/web/src/entities/fitness/ui/FitnessList.tsx
