+++ apps/web/src/entities/fitness/ui/FitnessListEditor.tsx
