+++ apps/web/src/entities/fitness/ui/FitnessListSearch.stories.tsx
