+++ apps/web/src/entities/fitness/ui/FitnessListSearch.tsx
