+++ apps/web/src/entities/fitness/ui/FitnessSearchModal.stories.tsx
