+++ apps/web/src/entities/fitness/ui/FitnessSearchModal.tsx
