+++ apps/web/src/entities/fitness/ui/SimpleFitnessItem.stories.tsx
