+++ apps/web/src/entities/fitness/ui/SimpleFitnessItem.tsx
@@ -1,7 +1,7 @@
 import type { Exercise } from 'fitness-struct'
 import { Suspense } from 'react'
 import { useFitnessFragment } from '@entities/fitness/api';
-import SetState from '@ui/Sets/SetState'
+import SetState from '@entities/set/ui/SetState'
 
 export interface SimpleFitnessItemProps {
   exerciseData: Exercise.Data
