+++ apps/web/src/entities/fitness/ui/SimpleFitnessList.stories.tsx
