+++ apps/web/src/entities/fitness/ui/SimpleFitnessList.tsx
