+++ apps/web/src/entities/schedule/ui/ScheduleDisplay.stories.tsx
