+++ apps/web/src/entities/schedule/ui/ScheduleDisplay.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from 'react';
 import { useGetExerciseListByScheduleId } from '@entities/exercise/api'
 import { useScheduleMenu } from '@entities/schedule/hooks';
 import MenuableAccordion from '@shared/ui/MenuableAccordion';
-import SimpleFitnessList from '@ui/Fitness/SimpleFitnessList'
+import SimpleFitnessList from '@entities/fitness/ui/SimpleFitnessList'
 
 export interface ScheduleDisplayProps {
   title: string,
