+++ apps/web/src/entities/set/ui/SetRow.tsx
