+++ apps/web/src/pages/:selectDate/schedule/create.tsx
@@ -3,7 +3,7 @@ import { useScheduleActions } from '@entities/schedule/hooks';
 import { useCreateScheduleWithExercisePlans } from '@features/exerciseManagement/api';
 import { LogEvent } from '@service/firebase';
 import usePageTracker from '@shared/hooks/usePageTracker';
-import FitnessListEditor from '@ui/Fitness/FitnessListEditor';
+import FitnessListEditor from '@entities/fitness/ui/FitnessListEditor';
 import { useAlert } from '@widgets/alert';
 import { useHeaderHandler } from '@widgets/header';
