+++ apps/web/src/pages/:selectDate/workout/:id.tsx
@@ -9,8 +9,8 @@ import { useScheduleHeaderMenu } from '@entities/schedule/hooks'
 import { ScheduleType } from '@entities/schedule/model/ScheduleType'
 import usePageTracker from '@shared/hooks/usePageTracker'
 import { dayjs } from '@shared/lib/utils'
-import ExerciseDataList from '@ui/ExerciseData/ExerciseDataList'
-import PresetNameInputDialog from '@ui/Preset/PresetNameInputDialog'
+import ExerciseDataList from '@entities/exercise/ui/ExerciseDataList'
+import PresetNameInputDialog from '@entities/exercisePreset/ui/PresetNameInputDialog'
 import { useAlert } from '@widgets/alert';
 import { BottomNaviArea } from '@widgets/bottomNavi'
 import { useHeaderHandler } from '@widgets/header'
