+++ apps/web/src/pages/Main.tsx
@@ -7,7 +7,7 @@ import { useScheduleActions } from '@entities/schedule/hooks'
 import usePageTracker from '@shared/hooks/usePageTracker'
 import MenuableAccordion from '@shared/ui/MenuableAccordion'
 import StateRender from '@shared/ui/StateRender'
-import ScheduleDisplay from '@ui/Schedule/ScheduleDisplay'
+import ScheduleDisplay from '@entities/schedule/ui/ScheduleDisplay'
 import { useBottomNavi } from '@widgets/bottomNavi'
 import { DateService } from '@widgets/calender'
 import { useHeaderHandler } from '@widgets/header'
